+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `Mentions` to `TopicEvent`
 - Added `Webhook.CreateUrl(webhookId, token)`
 - Added `ServerId` to `Member` and `MemberSummary` (this gets set in the behind the scenes of Guilded.NET, so bugs may occur)
+- Added `RemoveAsync`, `AddBanAsync`, `GetSocialLinkAsync`, etc. to `Member` and `MemberSummary`
 - Moved `GuildedWebhookClient` and `WebhookSkeleton` to `Guilded.Webhook` package
 - Fixed client hanging on reconnect
 - Removed/commented out `FormId` (might come back in the future, but not now)